Fixes install command (#433)
* Fixes install

* Remove shipit
