--- container/shipit/__init__.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*-
-from __future__ import absolute_import
-
-import logging
-
-logger = logging.getLogger(__name__)
-
