--- container/shipit/kubernetes/modules/kube_api.py
@@ -1,129 +1,0 @@
-import re
-import json
-
-
-class KubeAPI(object):
-
-    def __init__(self, module, target="kubectl"):
-        self.target = target
-        self.module = module
-
-    @staticmethod
-    def use_multiple_deployments(services):
-        '''
-        Inspect services and return True if the app supports multiple replica sets.
-
-        :param services: list of docker-compose service dicts
-        :return: bool
-        '''
-        multiple = True
-        for service in services:
-            if not service.get('ports'):
-                multiple = False
-            if service.get('volumes_from'):
-                multiple = False
-        return multiple
-
-    def call_api(self, cmd, data=None, check_rc=False, error_msg=None):
-        target_cmd = "%s %s" % (self.target, cmd)
-        rc, stdout, stderr = self.module.run_command(target_cmd, data=data)
-        logger.debug("Received rc: %s" % rc)
-        logger.debug("stdout:")
-        logger.debug(stdout)
-        logger.debug("stderr:")
-        logger.debug(stderr)
-
-        if check_rc and rc != 0:
-            self.module.fail_json(msg=error_msg, stderr=stderr, stdout=stdout)
-
-        return rc, stdout, stderr
-
-    def create_from_template(self, template=None, template_path=None):
-        if template_path:
-            logger.debug("Create from template %s" % template_path)
-            error_msg = "Error Creating %s" % template_path
-            cmd = "create -f %s" % template_path
-            rc, stdout, stderr = self.call_api(cmd, check_rc=True, error_msg=error_msg)
-            return stdout
-
-        if template:
-            logger.debug("Create from template:")
-            formatted_template = json.dumps(template, sort_keys=False, indent=4, separators=(',', ':'))
-            logger.debug(formatted_template)
-            cmd = "create -f -"
-            rc, stdout, stderr = self.call_api(cmd, data=formatted_template, check_rc=True,
-                                               error_msg="Error creating from template.")
-            return stdout
-
-    def replace_from_template(self, template=None, template_path=None):
-        if template_path:
-            logger.debug("Replace from template %s" % template_path)
-            cmd = "replace -f %s" % template_path
-            error_msg = "Error replacing %s" % template_path
-            rc, stdout, stderr = self.call_api(cmd, check_rc=True, error_msg=error_msg)
-            return stdout
-        if template:
-            logger.debug("Replace from template:")
-            formatted_template = json.dumps(template, sort_keys=False, indent=4, separators=(',', ':'))
-            logger.debug(formatted_template)
-            cmd = "replace -f -"
-            rc, stdout, stderr = self.call_api(cmd, data=formatted_template, check_rc=True,
-                                               error_msg="Error replacing from template")
-            return stdout
-
-    def delete_resource(self, type, name):
-        cmd = "delete %s/%s" % (type, name)
-        logger.debug("exec: %s" % cmd)
-        error_msg = "Error deleting %s/%s" % (type, name)
-        rc, stdout, stderr = self.call_api(cmd, check_rc=True, error_msg=error_msg)
-        return stdout
-
-    def get_resource(self, type, name):
-        result = None
-        cmd = "get %s/%s -o json" % (type, name)
-        logger.debug("exec: %s" % cmd)
-        rc, stdout, stderr = self.call_api(cmd)
-        if rc == 0:
-            try:
-                result = json.loads(str(stdout))
-            except Exception as exc:
-                self.module.fail_json("Error load json from kubectl output - %s" % str(exc))
-        elif rc != 0 and not re.search('not found', stderr):
-            self.module.fail_json(msg="Error getting %s/%s" % (type, name), stderr=stderr, stdout=stdout)
-        return result
-   
-    def set_context(self, context_name):
-        cmd = "use-context %s" % context_name
-        logger.debug("exec: %s" % cmd)
-        error_msg = "Error switching to context %s" % context_name
-        rc, stdout, stderr = self.call_api(cmd, check_rc=True, error_msg=error_msg)
-        return stdout
-
-    def set_project(self, project_name):
-        result = True
-        cmd = "project %s" % project_name
-        logger.debug("exec: %s" % cmd)
-        rc, stdout, stderr = self.call_api(cmd)
-        if rc != 0:
-            result = False
-            if not re.search('does not exist', stderr):
-                self.module.fail_json(msg="Error switching to project %s" % project_name, stderr=stderr, stdout=stdout)
-        return result
-
-    def create_project(self, project_name):
-        result = True
-        cmd = "new-project %s" % project_name
-        logger.debug("exec: %s" % cmd)
-        error_msg = "Error creating project %s" % project_name
-        self.call_api(cmd, check_rc=True, error_msg=error_msg)
-        return result
-
-    def get_deployment(self, deployment_name):
-        cmd = "%s deploy %s" % deployment_name
-        logger.debug("exec: %s" % cmd)
-        rc, stdout, stderr = self.call_api(cmd)
-        if rc != 0:
-            if not re.search('not found', stderr):
-                self.module.fail_json(msg="Error getting deployment state %s" % deployment_name, stderr=stderr,
-                                      stdout=stdout)
-        return stdout
